devtools/shared/webconsole/test/chrome/test_bug819670_getter_throws.html

<!DOCTYPE HTML>
<html lang="en">
  <head>
    <meta charset="utf8">
    <title>Test for Bug 819670 - Web console object inspection does not handle native getters throwing very well</title>
    <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="text/javascript" src="common.js"></script>
    <!-- Any copyright is dedicated to the Public Domain.
         - http://creativecommons.org/publicdomain/zero/1.0/ -->
  </head>
  <body>
    <p>Test for Bug 819670 - Web console object inspection does not handle native getters throwing very well</p>

    <script class="testbody" type="text/javascript">
      "use strict";

      SimpleTest.waitForExplicitFinish();
      addEventListener("load", startTest);

      async function startTest() {
        removeEventListener("load", startTest);
        const {state} = await attachConsoleToTab([]);
        const evaluationResponse = await state.webConsoleFront.evaluateJSAsync("document.__proto__");
        checkObject(evaluationResponse, {
          input: "document.__proto__",
          result: {
            type: "object",
            actor: /[a-z]/
          }
        });

        ok(!evaluationResponse.exception, "no eval exception");
        ok(!evaluationResponse.helperResult, "no helper result");

        const response = await evaluationResponse.result.getPrototypeAndProperties();
        ok(! response.error, "no response error");

        const props = response.ownProperties;
        ok(props, "response properties available");

        if (props) {
          // eslint-disable-next-line no-proto
          const expectedProps = Object.getOwnPropertyNames(document.__proto__);
          isDeeply(Object.keys(props), expectedProps, "Same own properties.");
        }

        await closeDebugger(state);
        SimpleTest.finish();
      }
    </script>
  </body>
</html>
